Enum values now have long type. Fixed template generation for enums.
git-svn-id: https://anonsvn.ncbi.nlm.nih.gov/repos/v1/trunk/c++@965 78c7ea69-d796-4a43-9a09-de51944f1b03

+++ include/serial/objistr.hpp
@@ -33,6 +33,10 @@
 *
 * ---------------------------------------------------------------------------
 * $Log$
+* Revision 1.28  1999/10/18 20:11:15  vasilche
+* Enum values now have long type.
+* Fixed template generation for enums.
+*
 * Revision 1.27  1999/10/04 16:22:08  vasilche
 * Fixed bug with old ASN.1 structures.
 *
@@ -150,7 +154,7 @@ public:
     
     // try to read enum value name, "" if none
     virtual string ReadEnumName(void);
-    virtual int ReadEnumValue(void);
+    virtual long ReadEnumValue(void);
 
     virtual TTypeInfo MapType(const string& name);
     void SetTypeMapper(CTypeMapper* typeMapper);
